Add files via upload
This project is a dice roller program written in Python. It allows the user to specify the number of dice and the number of sides on each dice, rolls the dice, and displays the results. The user can then choose to roll again or exit the program. The program makes use of the random module to generate random numbers for the rolls.